+++ test/fixture/tailwind.config.ts
@@ -1,7 +1,8 @@
 const colors = require('tailwindcss/colors')
 import {Config} from 'tailwindcss'
 
-const path = process.env.NODE_ENV == 'development' ? '../../tailwind/plugin' : 'digimarket-kit/plugin'
+const path = process.env.TW_PLUGIN_PATH || 'digimarket-kit/plugin'
+
 export default <Config>{
     content: [
         `../../src/components/**/*.{vue,js}`,
@@ -10,8 +11,11 @@ export default <Config>{
         `../../src/layouts/**/*.vue`, `pages/**/*.vue`,
         `../../src/composables/**/*.{js,ts}`,
         `../../src/plugins/**/*.{js,ts}`,
+        `pages/**/*.vue`,
+        `pages/*.vue`,
         `components/**/*.{vue,js}`,
-        `layouts/**/*.vue`, `pages/**/*.vue`,
+        `layouts/**/*.vue`,
+        `pages/**/*.vue`,
         `composables/**/*.{js,ts}`,
         `plugins/**/*.{js,ts}`,
         `App.{js,ts,vue}`,
@@ -19,10 +23,23 @@ export default <Config>{
         './node_modules/flowbite/**/*.js',
         './node_modules/tw-elements/dist/js/**/*.js'
     ],
+    darkMode: ['class', '[data-mode="dark"]'],
+    theme: {
+        colors: {
+            transparent: 'transparent',
+            secondary: '#008eb2',
+            error: '#f44336',
+            success: '#EE00EE',
+            alert: '#ff9800',
+            dark: '#212121',
+            primary: '#009000',
+            accent: '#343232',
+            light: '#EEE',
+        },
+    },
     plugins: [
-        require('flowbite/plugin'),
-        require('tw-elements/dist/plugin'),
-        require('digimarket-kit/plugin'),
-        // require(path)
+        // require('flowbite/plugin'),
+        // require('tw-elements/dist/plugin'),
+        require(path)
     ]
 }